+++ tests/update-dependencies.sh
@@ -1,41 +1,5 @@
 #!/bin/bash -ex
 
-test_ci() {
-  ../setup.sh $1 | tee setup.log
-  grep "$1" setup.log
-
-  circleci config validate
-  set +e
-  egrep -r '(my_module|MyModule)' * .circleci .gitignore
-  CHECK=$?
-  if [ $CHECK -eq 0 ]
-  then
-    set -e
-    exit 1
-  fi
-  set -e
-
-  # This module fails CS jobs currently so this is more informational.
-  if [ ! -z $1 ]
-  then
-    circleci.sh -e CIRCLE_PROJECT_REPONAME=node build --job run-code-sniffer || true
-    circleci.sh -e CIRCLE_PROJECT_REPONAME=node build --job run-unit-kernel-tests
-
-    circleci.sh -e CIRCLE_PROJECT_REPONAME=node build --job run-behat-tests | tee behat.log
-    # We need to skip colour codes
-    egrep "1 scenario \\(.*1 passed" behat.log
-
-    # Test that a PHP FATAL error properly fails the job.
-    git apply ../fixtures/behat-fail.patch
-
-    # circleci doesn't bubble the exit code from behat :(
-    circleci.sh -e CIRCLE_PROJECT_REPONAME=node build --job run-behat-tests | tee behat.log
-    grep -A9 'Behat tests failed' behat.log | tail -n 1 | grep '+ exit 1'
-
-    git reset --hard HEAD
-  fi
-}
-
 sudo apt-get update -y
 sudo apt-get install php5-cli -y
 EXPECTED_SIGNATURE=$(wget -q -O - https://composer.github.io/installer.sig)
@@ -64,8 +28,6 @@ fi
 
 git clone git@github.com:deviantintegral/drupal_tests_node_example.git node
 cd node
-git checkout 118b911
+git checkout drupal-node-86
 
-test_ci $1
-
-echo 'All tests have passed.'
+echo 'Test dependencies set up.'