+++ CHANGELOG.md
@@ -4,6 +4,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - YYYY-MM-DD
+
+### Changed
+
+* Updates to Drupal 8.6 #49
+  * Selenium has been upgraded to 3.14.0
+  * Full support for Functional Javascript tests
+  * Fixes running all tests in the unit/kernel test job
+  * `test-js.sh` has been renamed to `behat.sh`.
+  * `test-functional.sh` and `test-functional-js.sh` hooks have been added.
+
+## [0.3.4] - 2018-04-04
+
+### Changed
+
+* Update to Drupal 8.5 release #42
+* Update drupal and robo versions.
+
+### Fixed
+
+* Fix mink-selenium version conflict with Drupal 8.5 dev requirement
+* Fix composer-patches not triggering on first composer update #44
+
 ## [0.3.3] - 2018-02-28
 
 ### Fixed